+++ packages/markdown/package.js
@@ -8,6 +8,7 @@ Package.describe({
 Package.onUse(function (api) {
 	api.versionsFrom(['2.16', '3.0']);
 	api.use('templating');
+	api.use('reactive-var', 'client');
 	api.use("ecmascript", ['server', 'client']);
 
 	api.export('Markdown', ['server', 'client']);