+++ server/publications/settings.js
@@ -34,6 +34,7 @@ Meteor.publish('setting', () => {
         disableRegistration: 1,
         disableForgotPassword: 1,
         renderLinksAsPlainText: 1,
+        alwaysShowCodeAsText: 1,
         productName: 1,
         hideLogo: 1,
         hideCardCounterList: 1,